\name{summary_plots}
\alias{summary_plots}
\title{plot R^2, kOut, and kIn for a TRN model}
\description{plot the distribution of R^2, kOut, and kIn for a TRN model produced by fitTRN.}
\usage{summary_plots(trn)}
\arguments{
  \item{trn}{A TRN object, usually produced by fitTRN.}
}
\value{plot for the distribution of R^2, kOut, and kIn for trn.}
\author{Seth Ament}
