\name{tfbs.hsa.entrez}
\alias{tfbs.hsa.entrez}
\docType{data}
\title{
Model of Human Transcription Factor Binding Sites (Entrez ID's)
}
\description{
Intersection of cis-motifs from TRANSFAC, JASPAR, and Jolma et al. (2013) with digital genomic footprints from the ENCODE project.
}
\usage{data("tfbs.hsa.entrez")}
\format{
  The format is:
 num [1:18148, 1:1056] 1 1 1 1 1 1 1 1 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:18148] "163859" "57449" "114827" "8338" ...
  ..$ : chr [1:1056] "3142" "79192" "153572" "79191" ...
}