\name{tfbs.hsa.hgnc_symbol}
\alias{tfbs.hsa.hgnc_symbol}
\docType{data}
\title{Model of Human Transcription Factor Binding Sites (HGNC Gene Symbols)}
\description{
Intersection of cis-motifs from TRANSFAC, JASPAR, and Jolma et al. (2013) with digital genomics footprints from the ENCODE project.
}
\usage{data("tfbs.hsa.hgnc_symbol")}
\format{
  The format is:
 num [1:18067, 1:1021] 1 1 1 1 1 1 1 1 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:18067] "SDE2" "PLEKHG5" "FHAD1" "HIST2H2AC" ...
  ..$ : chr [1:1021] "HLX" "IRX1" "IRX2" "IRX3" ...
}