\name{tfbs.mmu.entrez}
\alias{tfbs.mmu.entrez}
\docType{data}
\title{
Model of Mouse Transcription Factor Binding Sites (Entrez ID's)
}
\description{
Intersection of cis-motifs from TRANSFAC, JASPAR, and Jolma et al. (2013) with digital genomics footprints from the mouseENCODE project.
}
\usage{data("tfbs.mmu.entrez")}
\format{
  The format is:
 num [1:18629, 1:878] 1 1 1 1 1 1 1 1 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:18629] "227377" "18582" "108735" "71884" ...
  ..$ : chr [1:878] "11545" "11614" "23957" "11622" ...
}