\name{tfbs.mmu.mgi_symbol}
\alias{tfbs.mmu.mgi_symbol}
\docType{data}
\title{
Model of Mouse Transcription Factor Binding Sites (MGI Gene Symbols)
}
\description{
Intersection of cis-motifs from TRANSFAC, JASPAR, and Jolma et al. (2013) with digital genomic footprints from the mouseENCODE project.
}
\usage{data("tfbs.mmu.mgi_symbol")}
\format{
  The format is:
 num [1:18627, 1:871] 1 1 1 1 1 1 1 1 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:18627] "Farp2" "Pde6d" "Sft2d2" "Chit1" ...
  ..$ : chr [1:871] "Parp1" "Nr0b1" "Nr0b2" "Ahr" ...
}