\name{trena-package}
\alias{trena-package}
\alias{trena}
\docType{package}
\title{Transcriptional Regulatory Network Analysis}
\description{Methods for Reconstructing Transcriptional Regulatory Networks}
\author{Seth Ament (Institute for Systems Biology) <seth.ament@systemsbiology.org>}
\references{Ament S.A. et al. in preparation}